Move repairstep to a custom command
People that have issues can run it manually
